new project built. Set up config, controller for register, models, repositories. Next time will work on html files.
