+++ DistributionCenters/src/main/java/com/cpan252/DistributionCenters/DistributionCenter.java
@@ -1,0 +1,11 @@
+package com.cpan252.DistributionCenters;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Cpan252Application {
+    public static void main(String[] args) {
+        SpringApplication.run(Cpan252Application.class, args);
+    }
+}
