+++ DistributionCenters/src/main/java/com/cpan252/DistributionCenters/config/WebConfig.java
@@ -1,0 +1,15 @@
+package com.cpan252.DistributionCenters.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addViewController("/").setViewName("/login");
+    }
+
+}
