+++ DistributionCenters/src/main/java/com/cpan252/DistributionCenters/model/DistributionCenter.java
@@ -1,0 +1,25 @@
+package com.cpan252.DistributionCenters.model;
+
+import java.math.BigDecimal;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.validation.constraints.NotBlank;
+
+@Entity
+public class DistributionCenter {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @NotBlank
+    private String name;
+    @NotBlank
+    private int itemsAvailable;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
+
+    // Add getters and setters here
+}
+
