+++ DistributionCenters/src/main/java/com/cpan252/DistributionCenters/repository/CenterRepository.java
@@ -1,0 +1,7 @@
+package com.cpan252.DistributionCenters.repository;
+
+import com.cpan252.model.DistributionCenter;
+import org.springframework.data.repository.CrudRepository;
+
+public interface CenterRepository extends CrudRepository<DistributionCenter, Long> {
+}
