+++ DistributionCenters/src/main/java/com/cpan252/DistributionCenters/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.cpan252.DistributionCenters.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import com.cpan252.DistributionCenters.model.User;
+
+@Repository
+public interface UserRepository extends CrudRepository<User, Long> {
+    User findByUsername(String username);
+}
