+++ DistributionCenters/src/main/resources/templates/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE HTML>
+<html >
+<head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+</head>
+<body>
+<h1>Tekken Reborn </h1>
+</body>
+</html>
