--- Final/src/main/java/com/cpan252/tekkenreborn/config/SecurityConfig.java
@@ -1,68 +1,0 @@
-package com.cpan252.tekkenreborn.config;
-import static org.springframework.boot.autoconfigure.security.servlet.PathRequest.toH2Console;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-import com.cpan252.tekkenreborn.model.User;
-import com.cpan252.tekkenreborn.repository.UserRepository;
-
-@Configuration
-@EnableMethodSecurity
-public class SecurityConfig {
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public UserDetailsService userDetailsService(UserRepository userRepository) {
-        return username -> {
-            User user = userRepository.findByUsername(username);
-            if (user != null) {
-                return user;
-            }
-            throw new UsernameNotFoundException(username);
-        };
-    }
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        return http
-                .authorizeHttpRequests()
-                .requestMatchers(toH2Console()).permitAll()
-                .requestMatchers("/form", "/clothinglist")
-                .hasRole("USER")
-                .anyRequest().permitAll()
-                .and()
-                .formLogin()
-                .loginPage("/login")
-                .defaultSuccessUrl("/form", true)
-                .and()
-                .logout()
-                .logoutSuccessUrl("/")
-
-                // Make H2-Console non-secured; for debug purposes
-                .and()
-                .csrf()
-                .ignoringRequestMatchers(toH2Console())
-                // Allow pages to be loaded in frames from the same origin; needed for
-                // H2-Console
-                .and()
-                .headers()
-                .frameOptions()
-                .sameOrigin()
-
-                .and()
-                .build();
-    }
-
-}
