--- Final/src/main/java/com/cpan252/tekkenreborn/controller/AddItemController.java
@@ -1,17 +1,0 @@
-package com.cpan252.tekkenreborn.controller;
-
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/addItem")
-public class AddItemController {
-    @GetMapping
-    @PreAuthorize("hasRole('ROLE_ADMIN') or hasRole('ROLE_EMPLOYEE')")
-    public String design() {
-        return "addItem";
-    }
-
-}
