--- Final/src/main/java/com/cpan252/tekkenreborn/controller/RegistrationController.java
@@ -1,32 +1,0 @@
-package com.cpan252.tekkenreborn.controller;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import com.cpan252.tekkenreborn.model.RegistrationForm;
-import com.cpan252.tekkenreborn.repository.UserRepository;
-
-@Controller
-@RequestMapping("/register")
-public class RegistrationController {
-    private UserRepository userRepository;
-    private PasswordEncoder passwordEncoder;
-
-    public RegistrationController(UserRepository userRepository, PasswordEncoder passwordEncoder) {
-        this.userRepository = userRepository;
-        this.passwordEncoder = passwordEncoder;
-    }
-
-    @GetMapping
-    public String showRegistrationForm() {
-        return "/register";
-    }
-
-    @PostMapping
-    public String registerUserAccount(RegistrationForm form) {
-        userRepository.save(form.toUser(passwordEncoder));
-        return "redirect:/";
-    }
-}
