--- Final/src/main/java/com/cpan252/tekkenreborn/model/ClothingDto.java
@@ -1,11 +1,0 @@
-package com.cpan252.tekkenreborn.model;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-
-public class ClothingDto {
-    private String name;
-}
