--- Final/src/main/java/com/cpan252/tekkenreborn/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.cpan252.tekkenreborn.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import com.cpan252.tekkenreborn.model.User;
-
-@Repository
-public interface UserRepository extends CrudRepository<User, Long> {
-    User findByUsername(String username);
-}
