--- Final/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-#also can be done by spring.datasource.name=tekkenreborn
-spring.datasource.url=jdbc:h2:mem:tekkenreborn
-spring.datasource.generate-unique-name=false
