--- Final/src/main/resources/application.yml
@@ -1,6 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:h2:mem:tekkenreborn
-    generate-unique-name: false
-server:
-  port: 8081
