--- Final/src/main/resources/schema.sql
@@ -1,7 +1,0 @@
-create table if not exists clothing (
-id int not null auto_increment,
-name varchar(50),
-year1 int not null,
-price decimal not null,
-brand_form varchar(50) not null
-);
