--- Final/src/main/resources/templates/clothinglist.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org">
-
-<head>
-    <title>Assignment 1</title>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.1/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
-</head>
-
-<body>
-<h1>Clothing List</h1>
-<div class="d-flex flex-row mb-3">
-    <div th:each="clothing: ${clothings}" class="form-horizontal">
-        <h4 th:text="${clothing.name}" />
-        <h4 th:text="${clothing.year1}"/>
-        <h4 th:text="|Price: ${clothing.price}|"/>
-        <h4 th:text="${clothing.brandFrom}"/>
-        </br>
-    </div>
-</div>
-<div class="d-flex flex-column bd-highlight justify-content-center align-items-center w-75 ms-5">
-    <form class="mt-3" method="post" th:object="${clothingDto}">
-        <div class="mb-3">
-            <label for="clothingName" class="form-label">Enter brand name:</label>
-            <input
-                    type="text"
-                    class="form-control"
-                    id="clothingName"
-                    th:field="*{name}"
-                    aria-describedby="clothingHelp"
-            />
-        </div>
-        <button class="btn btn-primary mt-5">
-            Search Brand
-        </button>
-    </form>
-</div>
-</body>
-</html>
