--- Final/src/main/resources/templates/login.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Login</title>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
-    <link
-            href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
-            rel="stylesheet"
-            integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
-            crossorigin="anonymous"
-    />
-</head>
-<body>
-<div class="container mt-5">
-    <div class="row justify-content-center">
-        <div class="col-md-6">
-            <div class="card">
-                <div class="card-header">
-                    <h4 class="card-title">Login</h4>
-                </div>
-                <div th:if="${error}">
-                    Unable to login. Check your username and password.
-                </div>
-                <div class="card-body">
-                    <form method="POST" th:action="@{/login}" id="loginForm">
-                        <div class="form-group">
-                            <label class="form-label" for="username">Username: </label>
-                            <input
-                                    class="form-control"
-                                    type="text"
-                                    name="username"
-                                    id="username"
-                            />
-                        </div>
-                        <div class="form-group">
-                            <label class="form-label" for="password">Password: </label>
-                            <input
-                                    class="form-control"
-                                    type="password"
-                                    name="password"
-                                    id="password"
-                            />
-                        </div>
-                        <button type="submit" class="btn btn-primary">Login</button>
-                    </form>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-</body>
-</html>
-
