--- Final/src/main/resources/templates/management.html
@@ -1,19 +1,0 @@
-<!-- tag::all[] -->
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Tekken Reborn</title>
-    <link
-            href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
-            rel="stylesheet"
-            integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
-            crossorigin="anonymous"
-    />
-</head>
-
-<body>
-<div class="d-flex flex-column bd-highlight justify-content-center align-items-center w-75 ms-5">
-    <h1>Management page</h1>
-</div>
-</body>
-</html>
