--- Final/src/main/resources/templates/register.html
@@ -1,40 +1,0 @@
-<!-- tag::all[] -->
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Tekken Reborn</title>
-    <link
-            href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
-            rel="stylesheet"
-            integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
-            crossorigin="anonymous"
-    />
-</head>
-
-<body>
-<div class="d-flex flex-column bd-highlight justify-content-center align-items-center w-75 ms-5">
-    <h1>Register</h1>
-
-    <form method="POST" th:action="@{/register}" id="registration">
-
-        <label class="form-label" for="username">Username: </label>
-        <input
-                class="form-control"
-                type="text"
-                name="username"
-                id="username"
-        /><br />
-
-        <label class="form-label" for="password">Password: </label>
-        <input
-                class="form-control"
-                type="password"
-                name="password"
-                id="password"
-        /><br />
-
-        <input class="btn btn-primary mt-5" type="submit" value="Register" />
-    </form>
-</div>
-</body>
-</html>
