--- Final/src/test/java/com/cpan252/tekkenreborn/HomeControllerTest.java
@@ -1,34 +1,0 @@
-package com.cpan252.tekkenreborn;
-
-import static org.hamcrest.Matchers.containsString;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.test.web.servlet.MockMvc;
-
-@WebMvcTest
-public class HomeControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    /**
-     * This test method tests that the home page is returned
-     * has correct content and status
-     * @throws Exception
-     */
-    @Test
-    public void testHomePage() throws Exception {
-        mockMvc.perform(get("/home"))
-                .andExpect(content().string(containsString("Tekken")))
-                .andExpect(status().isOk())
-                .andExpect(view().name("home"));
-
-    }
-
-}
